Create number_guess.py
With Python Skill Up on O'Reilly with Shaun Wassell. Random number selected and user has as many guesses as needed to guess the number. Currently set 1-10, but could go more, and could also limit the number of guesses, and adapt for even or odd numbers. Different ways to customize and keep learning with Python.